test(ha): split node detail contract coverage
